Change all Dxl data types to uint32_t

+++ niryo_one_driver/include/niryo_one_driver/dxl_motor_state.h
@@ -66,32 +66,32 @@ class DxlMotorState {
         void setId(uint8_t motor_id) { id = motor_id; } // allows to change tool motor easily
         
         // getters - state
-        uint32_t getPositionState()     { return state_pos; }
-        uint32_t getVelocityState()     { return state_vel; }
-        uint32_t getTorqueState()       { return state_torque; }
-        uint8_t getTemperatureState()   { return state_temperature; }
-        uint8_t getVoltageState()       { return state_voltage; }
-        uint8_t getHardwareErrorState() { return state_hw_error; }
+        uint32_t getPositionState()      { return state_pos; }
+        uint32_t getVelocityState()      { return state_vel; }
+        uint32_t getTorqueState()        { return state_torque; }
+        uint32_t getTemperatureState()   { return state_temperature; }
+        uint32_t getVoltageState()       { return state_voltage; }
+        uint32_t getHardwareErrorState() { return state_hw_error; }
 
         // setters - state
-        void setPositionState(uint32_t pos)     { state_pos = pos; }
-        void setVelocityState(uint32_t vel)     { state_vel = vel; }
-        void setTorqueState(uint32_t torque)    { state_torque = torque; }
-        void setTemperatureState(uint8_t temp)  { state_temperature = temp; }
-        void setVoltageState(uint8_t volt)      { state_voltage = volt; }
-        void setHardwareError(uint8_t hw_error) { state_hw_error = hw_error; }
+        void setPositionState(uint32_t pos)      { state_pos = pos; }
+        void setVelocityState(uint32_t vel)      { state_vel = vel; }
+        void setTorqueState(uint32_t torque)     { state_torque = torque; }
+        void setTemperatureState(uint32_t temp)  { state_temperature = temp; }
+        void setVoltageState(uint32_t volt)      { state_voltage = volt; }
+        void setHardwareError(uint32_t hw_error) { state_hw_error = hw_error; }
 
         // getters - command
         uint32_t getPositionCommand() { return cmd_pos; }
         uint32_t getVelocityCommand() { return cmd_vel; }
         uint32_t getTorqueCommand()   { return cmd_torque; }
-        uint8_t getLedCommand()       { return cmd_led; }
+        uint32_t getLedCommand()      { return cmd_led; }
 
         // setters - command
-        void setPositionCommand(uint16_t pos)  { cmd_pos = pos; }
-        void setVelocityCommand(uint16_t vel)  { cmd_vel = vel; }
-        void setTorqueCommand(uint16_t torque) { cmd_torque = torque; }
-        void setLedCommand(uint8_t led)        { cmd_led = led; }
+        void setPositionCommand(uint32_t pos)  { cmd_pos = pos; }
+        void setVelocityCommand(uint32_t vel)  { cmd_vel = vel; }
+        void setTorqueCommand(uint32_t torque) { cmd_torque = torque; }
+        void setLedCommand(uint32_t led)       { cmd_led = led; }
 
     private:
 
@@ -103,16 +103,16 @@ class DxlMotorState {
         uint32_t state_pos; 
         uint32_t state_vel;
         uint32_t state_torque;
-        uint8_t state_temperature;
-        uint8_t state_voltage;
-        uint8_t state_hw_error;
+        uint32_t state_temperature;
+        uint32_t state_voltage;
+        uint32_t state_hw_error;
 
         // write variables 
 
         uint32_t cmd_pos;
         uint32_t cmd_vel;
         uint32_t cmd_torque; 
-        uint8_t cmd_led;
+        uint32_t cmd_led;
 };
 
 #endif
